requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# YAML Configuration
PyYAML==6.0.1

# HTTP Requests
requests==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

